controllers and auth t,z
Controllers have been added to make new users to navigate and understand code easily, Authentication and authorization have been added for security purpose so that only the camp and reviews belongs to particular person can make changes.

+++ routes/reviews.js
@@ -3,32 +3,28 @@ const router = express.Router({mergeParams:true});
 // importing all models
 const Campground = require('../models/campground');
 const Review = require('../models/review');
-// JOI require its own Schema to validate DB data
-const {reviewSchema} = require('../schemas');
-
-const ExpressError = require('../utils/ExpressError');
 // Instead of nesting all routes in try cath we impliment one function and use it
 const catchAsync = require('../utils/catchAsync');
-
-const reviewCampground = (req, res, next) => {
-    const {error} = reviewSchema.validate(req.body)
-    if(error){
-        const msg = error.details.map(el => el.message).join(',');
-        throw new ExpressError(msg, 400);
-    }else{
-        next();
-    }
-}
+const {reviewCampground, authenticateUser} = require('../middlewares');
 
 
 // Adding Review
-
-router.post('/', reviewCampground, catchAsync( async (req, res) => {
+router.get('/', authenticateUser, async (req, res) => {
+    const campground = await Campground.findById(req.params.id);
+    if(!campground){
+        req.flash('error', 'Something Went Wrong!');
+        return res.redirect('/campgrounds');
+    }
+    res.redirect(`/campgrounds/${req.params.id}`);
+})
+router.post('/', authenticateUser, reviewCampground, catchAsync( async (req, res) => {
     const campground = await Campground.findById(req.params.id);
     const review = new Review(req.body.review);
     if(!review || !campground){
         req.flash('error', "Failed to add Review");
+        return res.redirect('/campgrounds');
     }
+    review.author = req.user._id;
     campground.reviews.push(review);
     await campground.save();
     await review.save();
@@ -40,8 +36,13 @@ router.post('/', reviewCampground, catchAsync( async (req, res) => {
 
 // Deleteing Review
 
-router.delete('/:reviewId', catchAsync( async (req, res) => {
+router.delete('/:reviewId', authenticateUser, catchAsync( async (req, res) => {
     const {id, reviewId} = req.params;
+    const reviewBelong = await Review.findById(reviewId);
+    if(!reviewBelong && !reviewBelong.author(req.user._id)){
+        req.flash('error', "Something Went Wrong");
+        return res.redirect('/campgrounds');
+    }
     const camp = await Campground.findByIdAndUpdate( id, {$pull : {reviews: reviewId} });
     const doc = await Review.findByIdAndDelete(reviewId);
     if(!doc || !camp){
